refactor: relocate save-load managers to systems
